Fix orchard release build hanging
For some reason `orchard` builds on revision `d0baa18f` were hanging indefinitely _only_ when building in release mode.

Did a git bisect to find a more recent revision that still builds, (but does not fail due to other upstream requirements)
and found `624a834` to be the first working one!