[#351] Write a Commit message Section for CONTRIBUTING.md
Closes #351

In order to make the repo's history consistent going forwards,
I've added some commit message guidelines.

The core of the suggestions were taken from this article:
https://cbea.ms/git-commit/#why-not-how

The main idea is that the repo is the source of truth and helps
both readers and maintainers to understand it better.

The rationale behind the title structure is that a clean
changelog can be manufacture by running `git log --oneline`
and a little bit of editing.

Small suggested typo fixes

Small suggested content changes

Included PR suggestions to shorten commits' first line